adding lesson3 and moving code from program to lesson3
